[HW-2-4-2] переделал stack
