+++ HW-2-4-2/HW-2-4-2.cpp
@@ -1,37 +1,17 @@
-#include "HW-2-4-2.h"
+#include <iostream>
+#include "LinkedListStack.h"
+#include "LinkedListStack.cpp"
 
-LinkedListStack::LinkedListStack() {
-    head = nullptr;
-}
-
-LinkedListStack::~LinkedListStack() {
-    while (head != nullptr) {
-        Node* temp = head;
-        head = head->next;
-        delete temp;
-    }
-}
-
-void LinkedListStack::push(int value) {
-    Node* newNode = new Node(value);
-    newNode->next = head;
-    head = newNode;
-}
-
-int LinkedListStack::pop() {
-    if (head == nullptr) {
-        throw "Stack is empty!";
-    }
-    int value = head->value;
-    Node* temp = head;
-    head = head->next;
-    delete temp;
-    return value;
-}
-
-int LinkedListStack::getTop() {
-    if (head == nullptr) {
-        throw "Stack is empty!";
-    }
-    return head->value;
+int main() {
+    LinkedListStack stack;
+    stack.push(1);
+    stack.push(2);
+    stack.push(3);
+    std::cout << "Top element: " << stack.getTop() << std::endl;
+    std::cout << "Pop element: " << stack.pop() << std::endl;
+    std::cout << "Pop element: " << stack.pop() << std::endl;
+    std::cout << "Top element: " << stack.getTop() << std::endl;
+    stack.push(4);
+    std::cout << "Top element: " << stack.getTop() << std::endl;
+    return 0;
 }
